+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to the "beta-bhere" library will be documented in this file.
 
 ---
 
+## [1.0.32] - 2023-10-19
+
+### Added
+- PanoramaLoader & MediaLoader video error handler
+- selfServiceSupportRequest api catchError
+
+---
+
 ## [1.0.31] - 2023-09-28
 
 ### Fixing